adjust verifygroupkind method
